feat: add ssr, link.isExternalFn props to useMdComponent hook
When useMdComponent is used in an SSR environment, a hydration mismatch will occur due to the use of window.location when rendering links.

This commit adds the `ssr` prop for use in SSR environment to prevent a hydration mismatch. However, external icons will no longer render and there is a need to use the `link.isExternalFn` fn to calculate external-ness in an ssr environment.