fix: add overlay to ensure calendar popover open state resets
previously, when clicking on the calendar button, the popover will open. If the button is clicked again, the open popover will be  blurred, but the button will set the open state to "open" again. However, since the popover was in its closing animation, the popover will not open again. This results in a state desynchronisation between the state and the render.

The easiest fix is to add a full screen overlay behind the popover when it is open, so any click outside will blur the calendar before allowing any other elements to be clicked