feat: fix Input prefill style regression, add `isPrefilled` prop to Datepicker and Select (#754)
This PR fixes the regression where `isPrefilled` state started to do nothing on the `Input` component due to downstream ChakraUI changes on their Input component.

This PR also adds the `isPrefilled` prop to the following components:
- `SingleSelect`
- `MultiSelect`
- `DatePicker`
- `DateRangePicker`

The prefilled styling follows the styling of the `Input` component.

See storybook for details.